requirements.txt
pytorch_lightning==1.9.0
Pillow==10.0.0
opencv-python==4.8.1.78
transformers
taming-transformers-rom1504 
tqdm
numpy
kornia
webdataset
omegaconf
einops
scikit-image
pymcubes==0.1.4
carvekit-colab
open3d
trimesh
easydict
nerfacc
imageio-ffmpeg==0.4.7
pycocotools
matplotlib
onnxruntime
onnx
connected-components-3d
scikit-learn==1.3.1
info-nce-pytorch
ninja==1.11.1.1
tensorboard==2.14.0
git+https://github.com/NVlabs/tiny-cuda-nn/#subdirectory=bindings/torch
git+https://github.com/openai/CLIP.git
git+https://github.com/facebookresearch/segment-anything.git
